Login page, deleted two doc that weren't being used.
Editing the login to allow for the cursor to change to a hand. In
addition, I'm deling the forgotPassword.js and the resetPassword.js
becuase we did not add those pages. However, I do still have a copy of
the js if we decided that we wanted to add them.